+++ src/java/com/tasktoys/archelon/service/DiscussionService.java
@@ -4,7 +4,6 @@
 package com.tasktoys.archelon.service;
 
 import com.tasktoys.archelon.data.entity.Discussion;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -15,11 +14,16 @@ import java.util.Map;
 public interface DiscussionService {
     
     public List<Discussion> getNewestDiscussionList(int n);
-    public List<Discussion> getDiscussionListBefor(BigInteger id, int n);
+    public List<Discussion> getDiscussionListBefor(Long id, int n);
     public List<Discussion> getNewestDiscussionListByMainCategory(int n, int main_id);
-    public List<Discussion> getDiscussionListWithMainCategoryBefor(BigInteger id, int n, int main_id);
+    public List<Discussion> getDiscussionListWithMainCategoryBefor(Long id, int n, int main_id);
     public List<Discussion> getNewestDiscussionListBySubCategory(int n, int main_id, int sub_id);
-    public List<Discussion> getDiscussionListWithSubCategoryBefor(BigInteger id, int n, int main_id, int sub_id);
+    public List<Discussion> getDiscussionListWithSubCategoryBefor(Long id, int n, int main_id, int sub_id);
     
+    /**
+     * 
+     * @param dls
+     * @return 
+     */
     public List<Map<String, String>> replaceAuthorIDToAurthorName(List<Discussion> dls);
 }
